Polyfills use ESM
- Use "mjs" extension to prevent node errors
- Export polyfills as self-invoking functions
- Update entrypoint references
