--- packages/uswds-core/src/js/polyfills/number-is-nan.js
@@ -1,6 +1,0 @@
-Number.isNaN =
-  Number.isNaN ||
-  function isNaN(input) {
-    // eslint-disable-next-line no-self-compare
-    return typeof input === "number" && input !== input;
-  };
