+++ packages/uswds-core/src/js/polyfills/number-is-nan.mjs
@@ -1,0 +1,8 @@
+export default (function () {
+  Number.isNaN =
+    Number.isNaN ||
+    function isNaN(input) {
+      // eslint-disable-next-line no-self-compare
+      return typeof input === "number" && input !== input;
+    };
+})();
